booksからprogresses,target_settingsファイルに分割 #9
